requirements.txt
numpy
Pillow
opencv-python
transformers
huggingface-hub
torch

opencv-python
pandas
scikit-learn
scikit-image
Pillow
tqdm
pytoshop
psd_tools

onnx
onnxruntime-gpu
huggingface_hub
segment_anything
requests
numpy
einops
torch